fix(convex): add missing title field to threads table schema
- Add @mastra/core/storage/constants export path for Node.js-free schema access
- Rewrite Convex schema to dynamically build from TABLE_SCHEMAS
- Remove redundant server/schema.ts
- Add schema sync tests to prevent future drift

Fixes #11319
